refactor: CitiesCache is class now
